# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qorix Markets API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  fullName: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  fullName: string;
  isAdmin: boolean;
  referralCode: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Wallet {
  id: number;
  userId: number;
  mainBalance: number;
  tradingBalance: number;
  profitBalance: number;
  updatedAt: string;
}

export interface DepositBody {
  amount: number;
  /** @nullable */
  txHash?: string | null;
}

export interface WithdrawBody {
  amount: number;
  walletAddress: string;
}

export interface TransferBody {
  amount: number;
}

export interface Transaction {
  id: number;
  userId: number;
  type: string;
  amount: number;
  status: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface TransactionList {
  data: Transaction[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Investment {
  id: number;
  userId: number;
  amount: number;
  riskLevel: string;
  isActive: boolean;
  autoCompound: boolean;
  totalProfit: number;
  dailyProfit: number;
  drawdown: number;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  stoppedAt?: string | null;
}

export interface StartInvestmentBody {
  amount: number;
  riskLevel: string;
}

export interface CompoundingBody {
  autoCompound: boolean;
}

export interface Trade {
  id: number;
  symbol: string;
  direction: string;
  entryPrice: number;
  exitPrice: number;
  profit: number;
  profitPercent: number;
  executedAt: string;
}

export interface Referral {
  referralCode: string;
  totalReferred: number;
  activeReferrals: number;
  totalEarned: number;
  monthlyEarnings: number;
}

export interface ReferredUser {
  id: number;
  fullName: string;
  email: string;
  investmentAmount: number;
  isActive: boolean;
  joinedAt: string;
}

export interface DashboardSummary {
  totalBalance: number;
  dailyProfitLoss: number;
  dailyProfitPercent: number;
  activeInvestment: number;
  totalProfit: number;
  profitBalance: number;
  tradingBalance: number;
  nextPayoutDate: string;
  daysUntilPayout: number;
  /** @nullable */
  riskLevel?: string | null;
  isTrading: boolean;
}

export interface EquityPoint {
  date: string;
  equity: number;
  profit: number;
}

export interface AdminStats {
  totalUsers: number;
  activeInvestors: number;
  totalAUM: number;
  totalProfitPaid: number;
  pendingWithdrawals: number;
  pendingWithdrawalAmount: number;
  dailyProfitPercent: number;
}

export interface SetDailyProfitBody {
  profitPercent: number;
}

export interface AdminUser {
  id: number;
  email: string;
  fullName: string;
  isAdmin: boolean;
  mainBalance: number;
  tradingBalance: number;
  profitBalance: number;
  investmentAmount: number;
  isTrading: boolean;
  referralCode: string;
  createdAt: string;
}

export interface AdminUserList {
  data: AdminUser[];
  total: number;
  page: number;
  totalPages: number;
}

export interface WithdrawalRequest {
  id: number;
  userId: number;
  userEmail: string;
  userFullName: string;
  amount: number;
  walletAddress: string;
  status: string;
  requestedAt: string;
  /** @nullable */
  processedAt?: string | null;
}

export type GetTransactionsParams = {
  page?: number;
  limit?: number;
  type?: string;
};

export type GetTradesParams = {
  limit?: number;
};

export type GetEquityChartParams = {
  days?: number;
};

export type GetAdminUsersParams = {
  page?: number;
  limit?: number;
};