# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Qorix Markets API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # WALLET
...
  /admin/profit:
    post:
      operationId: setDailyProfit
      tags: [admin]
      summary: Set daily profit percentage for all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDailyProfitBody"
      responses:
        "200":
          description: Profit set and distributed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
...
  /admin/withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve a withdrawal request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
  /admin/withdrawals/{id}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject a withdrawal request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # AUTH
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        referralCode:
          type: ["string", "null"]
      required:
        - email
        - password
        - fullName
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        isAdmin:
          type: boolean
        referralCode:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - fullName
        - isAdmin
        - referralCode
        - createdAt

    # WALLET
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        activeInvestors:
          type: integer
        totalAUM:
          type: number
        totalProfitPaid:
          type: number
        pendingWithdrawals:
          type: integer
        pendingWithdrawalAmount:
          type: number
        dailyProfitPercent:
          type: number
      required:
        - totalUsers
        - activeInvestors
        - totalAUM
        - totalProfitPaid
        - pendingWithdrawals
        - pendingWithdrawalAmount
        - dailyProfitPercent
    SetDailyProfitBody:
      type: object
      properties:
        profitPercent:
          type: number
      required:
        - profitPercent
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        isAdmin:
          type: boolean
        mainBalance:
          type: number
        tradingBalance:
          type: number
        profitBalance:
          type: number
        investmentAmount:
          type: number
        isTrading:
          type: boolean
        referralCode:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - fullName
        - isAdmin
        - mainBalance
        - tradingBalance
        - profitBalance
        - investmentAmount
        - isTrading
        - referralCode
        - createdAt
    AdminUserList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - totalPages
    WithdrawalRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userEmail:
          type: string
        userFullName:
          type: string
        amount:
          type: number
        walletAddress:
          type: string
        status:
          type: string
        requestedAt:
          type: string
        processedAt:
          type: ["string", "null"]
      required:
        - id
        - userId
        - userEmail
        - userFullName
        - amount
        - walletAddress
        - status
        - requestedAt