# lib/api-zod/src/generated/types/withdrawalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qorix Markets API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WithdrawalRequest {
  id: number;
  userId: number;
  userEmail: string;
  userFullName: string;
  amount: number;
  walletAddress: string;
  status: string;
  requestedAt: string;
  /** @nullable */
  processedAt?: string | null;
}